+++ jwt-server/spring/jwt-spring-grpc-ecosystem/src/main/java/org/entur/jwt/spring/grpc/ecosystem/GrpcEcosystemAutoConfiguration.java
@@ -1,7 +1,5 @@
 package org.entur.jwt.spring.grpc.ecosystem;
 
-import io.grpc.MethodDescriptor;
-import io.grpc.ServerCall;
 import net.devh.boot.grpc.server.autoconfigure.GrpcServerSecurityAutoConfiguration;
 import net.devh.boot.grpc.server.security.authentication.AnonymousAuthenticationReader;
 import net.devh.boot.grpc.server.security.authentication.CompositeGrpcAuthenticationReader;
@@ -42,7 +40,6 @@ import org.springframework.security.authentication.AnonymousAuthenticationProvid
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.AuthenticationProvider;
 import org.springframework.security.authentication.ProviderManager;
-import org.springframework.security.core.Authentication;
 import org.springframework.security.core.authority.AuthorityUtils;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.oauth2.core.DelegatingOAuth2TokenValidator;
@@ -120,29 +117,7 @@ public class GrpcEcosystemAutoConfiguration {
             }
 
             if(!serviceNameMethodName.isEmpty()) {
-                defaultAccessPredicate = new AccessPredicate() {
-                    @Override
-                    public boolean test(Authentication authentication, ServerCall<?, ?> serverCall) {
-                        MethodDescriptor<?, ?> method = serverCall.getMethodDescriptor();
-
-                        String lowerCaseServiceName = method.getServiceName().toLowerCase();
-                        List<String> methodNames = serviceNameMethodName.get(lowerCaseServiceName);
-                        if (methodNames != null) {
-                            if (methodNames.contains("*")) {
-                                return true;
-                            }
-
-                            String lowerCaseBareMethodName = method.getBareMethodName().toLowerCase();
-                            String lowerCaseFullMethodName = method.getFullMethodName().toLowerCase();
-
-                            if (methodNames.contains(lowerCaseBareMethodName) || methodNames.contains(lowerCaseFullMethodName)) {
-                                return true;
-                            }
-                        }
-
-                        return AccessPredicate.fullyAuthenticated().test(authentication, serverCall);
-                    }
-                };
+                defaultAccessPredicate = new GrpcAnonymousAccessPredicate(serviceNameMethodName);
             }
         }
 
@@ -160,9 +135,10 @@ public class GrpcEcosystemAutoConfiguration {
     }
 
     @Bean
-    @ConditionalOnMissingBean(IssuerAuthenticationProvider.class)
-    public IssuerAuthenticationProvider issuerAuthenticationProvider(JwkSourceMap jwkSourceMap, List<JwtAuthorityEnricher> jwtAuthorityEnrichers, List<OAuth2TokenValidator<Jwt>> jwtValidators, Flavours flavours) {
-        log.info("Configure Grpc security");
+    @ConditionalOnMissingBean(AuthenticationManager.class)
+    public AuthenticationManager authenticationManager(JwkSourceMap jwkSourceMap, List<JwtAuthorityEnricher> jwtAuthorityEnrichers, List<OAuth2TokenValidator<Jwt>> jwtValidators, Flavours flavours) {
+
+        log.info("Configure GRPC security");
 
         if (flavours.isEnabled()) {
             List<JwtAuthorityEnricher> enrichers = new ArrayList<>(jwtAuthorityEnrichers);
@@ -180,26 +156,16 @@ public class GrpcEcosystemAutoConfiguration {
             jwtAuthorityEnrichers = enrichers;
         }
 
-        return IssuerAuthenticationProvider.newBuilder()
+        IssuerAuthenticationProvider provider = IssuerAuthenticationProvider.newBuilder()
                 .withJwkSourceMap(jwkSourceMap)
                 .withJwtValidators(jwtValidators)
                 .withJwtAuthorityEnrichers(jwtAuthorityEnrichers)
                 .build();
-    }
 
-    private DelegatingOAuth2TokenValidator<Jwt> getJwtValidators(String issuer, List<OAuth2TokenValidator<Jwt>> jwtValidators) {
-        List<OAuth2TokenValidator<Jwt>> validators = new ArrayList<>();
-        validators.add(new JwtIssuerValidator(issuer)); // this check is implicit, but lets add it regardless
-        validators.addAll(jwtValidators);
-        DelegatingOAuth2TokenValidator<Jwt> validator = new DelegatingOAuth2TokenValidator<>(validators);
-        return validator;
-    }
-
-    @Bean
-    @ConditionalOnMissingBean(UserDetailsService.class)
-    public AuthenticationManager authenticationManager(IssuerAuthenticationProvider provider) {
         final List<AuthenticationProvider> providers = new ArrayList<>();
         providers.add(provider);
+
+        // TODO disable this if no anon access?
         providers.add(new AnonymousAuthenticationProvider("key"));
         return new ProviderManager(providers);
     }
